pyproject.toml
[project]
name = "red-runner"
version = "0.1.0"
description = "Minimal guardrail reconnaissance runner — load attack catalogs, fire multi-turn probes, score with timing fingerprint."
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
red-runner = "recon.cli:main"
